+++ tests/ds/test_esa_cci_odp.py
@@ -656,3 +656,18 @@ class MakeLocalTest(unittest.TestCase):
         ds = data_source.make_local(random_string)
         self.assertIsNotNone(ds)
         local_data_store.remove_data_source(f"local.{random_string}")
+
+
+@unittest.skip(reason='Used for debugging issue with duplicate dsr_ids')
+class NoDuplicatesTest(unittest.TestCase):
+    def test_for_duplicates_in_drs_ids(self):
+        data_store = EsaCciOdpDataStore()
+        data_sets = data_store.query()
+        ids = []
+        for dataset in data_sets:
+            ids.append(dataset.id)
+        if len(ids) == len(set(ids)):
+            contains_duplicates = False
+        else:
+            contains_duplicates = True
+        self.assertFalse(contains_duplicates)